Here, I've introduced Maven to the project, and cleaned up the packaging a little bit.
Maven is a build system, which lets you run `mvn package` to build the
entire project. This is what's used professionally, along with Gradle,
though I prefer Maven. It lets you do things such as compile your
entire source in one go, rather than one file at a time, and can track
dependencies. This is better than tracking things through an IDE.

The reason this is useful is because it allows you to then have a
server building your software for you. If you have a bunch of tests
built in to your source, the server can run the tests through Maven,
then, if they all pass, bundle everything up, and upload it to your
website as a released version. This is called continuous integration,
and you'll see it all over the place. You may have heard of
TravisCI. This is what Travis does. You might also encounter Jenkins,
which is another one.

The package naming scheme was changed slightly to instead be
`com.projectzeus.App`. This has no functional effect, however is
convention. By convention, the Java package should start with your
company website in reverse - `com.somedomain.xxx`, followed by the
name of the project you're working on. If you don't have a domain, or
a company, `com.projectname` is the other alternative. App is the
conventional entry point for command line apps. By doing this, other
people will know where to find things, and it keeps your source code
organised a little better.