Here, I set up an interface, Enemy, which details all of the methods enemies will need.
Interfaces are a way by which Java lets us do polymorphism. What this
means is that we don't actually care what class we have, only that it
can do certain things. Doing this, we can treat different classes the
same. In our enemies example, we can have a Heavy enemy, and a Light
enemy, and we don't really care which we've got, only that it's an
enemy.

In this case, all enemies have a name, and all enemies will have some
kind of attack damage they can do, different to each enemy. So, we can
now have as many enemy classes as we want, each implementing this
interface, and we can treat them all the same. Later on, if we decide
to add a new enemy, we don't need to update everywhere any more, and
if we decide enemies can do more, we can just add new stuff to the
interface.