Here I actually create the enemy classes, and add a couple of methods I missed the first time.
Note that the enemies are mostly the same, other than the constants at
the top. This means we could have used this using subclassing, rather
than interfaces, but this is fine. The way it's done here is more
elegant, and allows for much more flexibility - if ever damage needed
to be calculated in a different way for one specific enemy, for
example, it could be.