Here we introduce an EnemyFactory. This will be how we get enemies in the game logic itself.
The rationalle here is that, by using a factory, we don't have to hard
code what enemies the game can handle all over the logic. Every time
we need a new enemy, we request the factory for one, and only it knows
the complete list of enemies.

This way, if there are, say, 10 different places where enemies can be
spawned in the code, we only need to update one piece of code whenever
we add a new enemy class, as opposed to 10. If later on we need
different logic, if, for example, different enemies appear in
different proportions depending on their strength and the player
progression, that can all be done in our factory, keeping the external
interface nice and clean. We still only need EnemyFactory.getEnemy().