Here we hook everything up using our new factory and interface, as well as correct a few small mistakes that were made along the way in syntax and logic.
After having done this, the main class is significantly cleaner, and
easier to read, with all the irrelavent details on enemies abstracted
away into a clean factory and interface API in the enemy package. By
abstracting details away like this, the code is made far neater, more
easily readable, and is less prone to errors and bugs.